#4566 landed after this branch's base. It conflicts in clahe.py's `clip_limit` Args entry (main's
entry is the later, fuller one; this keeps it and folds in the per-sample sentence) and it added
`test_wart_random_clahe_applies_the_first_clip_limit_to_the_batch_4572`, a strict pin asserting the
batch-wide clip limit this PR replaces, plus prose in clahe.py and intensity/base.py describing it.
The pin is inverted into two convention pins and the two prose sites are rewritten.

+++ kornia/morphology/morphology.py
@@ -112,9 +112,12 @@ def dilation(
     elif engine == "convolution":
         B, C, H, W = tensor.size()
         h_pad, w_pad = output.shape[-2:]
-        reshape_kernel = _neight2channels_like_kernel(kernel)
+        reshape_kernel = _neight2channels_like_kernel(kernel).to(dtype=output.dtype)
         output, _ = F.conv2d(
-            output.view(B * C, 1, h_pad, w_pad), reshape_kernel, padding=0, bias=neighborhood.view(-1).flip(0)
+            output.view(B * C, 1, h_pad, w_pad),
+            reshape_kernel,
+            padding=0,
+            bias=neighborhood.view(-1).flip(0).to(dtype=output.dtype),
         ).max(dim=1)
         output = output.view(B, C, H, W)
     else:
@@ -205,9 +208,12 @@ def erosion(
     elif engine == "convolution":
         B, C, H, W = tensor.size()
         Hpad, Wpad = output.shape[-2:]
-        reshape_kernel = _neight2channels_like_kernel(kernel)
+        reshape_kernel = _neight2channels_like_kernel(kernel).to(dtype=output.dtype)
         output, _ = F.conv2d(
-            output.view(B * C, 1, Hpad, Wpad), reshape_kernel, padding=0, bias=-neighborhood.view(-1)
+            output.view(B * C, 1, Hpad, Wpad),
+            reshape_kernel,
+            padding=0,
+            bias=-neighborhood.view(-1).to(dtype=output.dtype),
         ).min(dim=1)
         output = output.view(B, C, H, W)
     else:
